# docs/openapi.yaml
openapi: 3.1.0
info:
  title: APEX Recommendation API
  description: 'Production-grade AI recommendation engine with a 6-model ensemble
    (SASRec, KAN, LightGCN, Diffusion, Quantum-Fluid, Hyperbolic). Implements Doubly
    Robust IPS weight optimization, differential privacy, and adaptive 3-tier serving
    (GPU / ONNX CPU / FAISS lite).


    **Interactive docs:** `/docs` (Swagger UI) · `/redoc` (ReDoc)


    **Full API reference:** [docs/API_REFERENCE.md](https://github.com/pavanpajjuri/Movie-Recommendation-System/blob/main/docs/API_REFERENCE.md)'
  contact:
    name: APEX Project
    url: https://github.com/pavanpajjuri/Movie-Recommendation-System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.0
paths:
...
  /v1/catalog/preview:
    post:
      tags:
      - Catalog
      summary: Preview Catalog
      operationId: preview_catalog_v1_catalog_preview_post
      parameters:
      - name: X-Nova-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Nova-Api-Key
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Id
      - name: X-Catalog-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Catalog-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogPreviewRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/catalog/upload:
    post:
      tags:
      - Catalog
      summary: Upload Catalog
      operationId: upload_catalog_v1_catalog_upload_post
      parameters:
      - name: X-Nova-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Nova-Api-Key
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Id
      - name: X-Catalog-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Catalog-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogUploadRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /movies:
    get:
      tags:
      - Browse
      summary: List Movies
      operationId: list_movies_movies_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          description: Maximum movies to return
          default: 100
          title: Limit
        description: Maximum movies to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Offset for pagination
          default: 0
          title: Offset
        description: Offset for pagination
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /movies/titles:
    get:
      tags:
      - Browse
      summary: Get All Titles
      operationId: get_all_titles_movies_titles_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100000
          minimum: 1
          description: Maximum number of titles to return
          default: 100000
          title: Limit
        description: Maximum number of titles to return
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/semantic-twins/id/{movie_id}:
    get:
      tags:
      - Browse
      summary: Semantic Twin By Id
      operationId: semantic_twin_by_id_v1_semantic_twins_id__movie_id__get
      parameters:
      - name: movie_id
        in: path
        required: true
        schema:
          type: integer
          title: Movie Id
      - name: X-Nova-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Nova-Api-Key
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Id
      - name: X-Catalog-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Catalog-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CatalogPreviewRequest:
      properties:
        filename:
          type: string
          title: Filename
          default: catalog.csv
        csv_text:
          type: string
          title: Csv Text
        column_mapping:
          additionalProperties:
            type: string
          type: object
          title: Column Mapping
        sample_size:
          type: integer
          title: Sample Size
          default: 20
      type: object
      required:
      - csv_text
      title: CatalogPreviewRequest
      description: CSV catalog preview request from the Nova Console.
    CatalogUploadRequest:
      properties:
        filename:
          type: string
          title: Filename
          default: catalog.csv
        csv_text:
          type: string
          title: Csv Text
        column_mapping:
          additionalProperties:
            type: string
          type: object
          title: Column Mapping
        sample_size:
          type: integer
          title: Sample Size
          default: 20
      type: object
      required:
      - csv_text
      title: CatalogUploadRequest
      description: Persisted catalog upload request.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
servers:
- url: http://localhost:8000
  description: Local development
- url: https://your-api.onrender.com
  description: Production (Render free tier — Tier 3)